requirements-xtts-train.txt
coqui-tts
trainer
soundfile
torchaudio
numpy
